Fix event listener example
In the example callback params were on the wrong places. According to
the documentation
(https://panoramisk.readthedocs.io/en/latest/manager.html#panoramisk.Manager.register_event)
the first argument should be manager and the second one should be the
event message, while in the example the first argument is called event
(while being actually manager) and the second one is called manager,
while actually being the event message.

This commit fixes that misleading example according to the documentation